+++ internal/js_parser/js_parser.go
@@ -10285,6 +10285,7 @@ func (p *parser) visitClass(nameScopeLoc logger.Loc, class *js_ast.Class) js_ast
 		oldIsThisCaptured := p.fnOnlyDataVisit.isThisNested
 		oldThis := p.fnOnlyDataVisit.thisClassStaticRef
 		oldShouldLowerSuper := p.fnOrArrowDataVisit.shouldLowerSuper
+		p.fnOrArrowDataVisit.shouldLowerSuper = false
 		p.fnOnlyDataVisit.isThisNested = true
 		p.fnOnlyDataVisit.isNewTargetAllowed = true
 		p.fnOnlyDataVisit.thisClassStaticRef = nil