+++ scripts/end-to-end-tests.js
@@ -4754,8 +4754,24 @@
             Derived.__proto__ = { foo() { log.push(15) } };
             Derived.bar();
 
+            Base = class { foo() { return 'bar' } };
+            Derived = class extends Base { async x() { return class { [super.foo()] = 123 } } };
+            if (new (await new Derived().x())().bar === 123) log.push(16);
+
+            Base = class { foo() { return 'bar' } };
+            Derived = class extends Base { x = async () => class { [super.foo()] = 123 } };
+            if (new (await new Derived().x())().bar === 123) log.push(17);
+
+            Base = class { static foo() { return 'bar' } };
+            Derived = class extends Base { static async x() { return class { [super.foo()] = 123 } } };
+            if (new (await Derived.x())().bar === 123) log.push(18);
+
+            Base = class { static foo() { return 'bar' } };
+            Derived = class extends Base { static x = async () => class { [super.foo()] = 123 } };
+            if (new (await Derived.x())().bar === 123) log.push(19);
+
             const observed = log.join(',');
-            const expected = '1,2,3,4,5,6,7,8,9,10,11,12,13,14,15';
+            const expected = '1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19';
             if (observed !== expected) throw 'fail: ' + observed + ' != ' + expected;
           }
         `,